Clicker first floor + upgrader

+++ Assets/Scripts/Resources/ResourceBank.cs
@@ -1,11 +1,11 @@
 public class ResourceBank
 {
-    private long _money;
+    private float _money;
     private int _multiplier;
-    public long Money { get => _money; set => _money = value; }
+    public float Money { get => _money; set => _money = value; }
     public int Multiplier { get => _multiplier; set => _multiplier = value; }
 
-    public ResourceBank(int money, int multiplier)
+    public ResourceBank(float money, int multiplier)
     {
         Money = money;
         Multiplier = multiplier;
